+++ space-d/src/main/java/com/dnd/spaced/core/account/infrastructure/persistence/AccountQuerydslRepository.java
@@ -1,4 +1,4 @@
-package com.dnd.spaced.core.account.infrastructure.repository;
+package com.dnd.spaced.core.account.infrastructure.persistence;
 
 import static com.dnd.spaced.core.account.domain.QAccount.account;
