+++ space-d/src/main/java/com/dnd/spaced/core/account/infrastructure/persistence/NicknameMetadataQuerydslRepository.java
@@ -1,4 +1,4 @@
-package com.dnd.spaced.core.account.infrastructure.repository;
+package com.dnd.spaced.core.account.infrastructure.persistence;
 
 import static com.dnd.spaced.core.account.domain.QNicknameMetadata.nicknameMetadata;
